% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.snpcov.matrix.R
\name{create.snpcov.matrix}
\alias{create.snpcov.matrix}
\title{create a matrix from pairwise data}
\usage{
create.snpcov.matrix(snpcov, snps)
}
\arguments{
\item{snpcov}{pairwise covariance matrix}

\item{snps}{a vector of snps to subset}
}
\description{
create a matrix from pairwise data
}

