% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.zscore.R
\name{impute.zscore}
\alias{impute.zscore}
\title{impute the z score}
\usage{
impute.zscore(gene.name, gwas, db, snpcov)
}
\arguments{
\item{gene.name}{gene name}

\item{gwas}{gwas data frame}

\item{db}{data frame containing the prediction models}

\item{snpcov}{data frame containing the pairwise snp covariances}
}
\description{
impute the z score
}
