% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaxcan.R
\name{metaxcan}
\alias{metaxcan}
\title{metaxcan main function}
\usage{
metaxcan(gwas.file, db.file, snpcov.file, genes = NA,
  return.snpinfo = FALSE, ncores = 1)
}
\arguments{
\item{gwas.file}{gwas summary stats file name}

\item{db.file}{prediction model db file name}

\item{snpcov.file}{pairwise snp-covariance file name}
}
\description{
metaxcan main function
}
