% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wts.R
\name{plot.wts}
\alias{plot.wts}
\title{plot weights of predictors and highlight a snp of interest}
\usage{
\method{plot}{wts}(snp, gene, db.file)
}
\arguments{
\item{snp}{snp id}

\item{gene}{gene id}

\item{db.file}{db file name with full path}
}
\description{
plot weights of predictors and highlight a snp of interest
}
